Aims to improve the accuracy of restoring the scroll position in draft mode on reload (see #16):
Before this non-applied (and abandoned-halfway-through) change, the slide position was/is only restored with slide-level granularity: Whichever slide's top edge is visible and closest to the viewport's top edge is being scrolled to (i.e. aligned to the viewport's top edge) on refresh.

That's ok, but not ideal! This implementation aims to bring this to pixel-level precision, but that doesn't work reliably and appears to be kind of much, both in complexity and timing sensitivity. Ugh.